Add support for log directory
One of our miners got stuck in an endless crash-restart loop.
Having the logs inside the container made it hard to check what
was going on.

By mounting the log folder externally, the log files are easily
accessible.

If no log directory is given, logs will be kept inside the container
just as before.